Move replace_esi_tags to _utils
