+++ armstrong/esi/_utils.py
@@ -2,7 +2,10 @@ from email.utils import parsedate
 import time
 
 from django.utils.cache import cc_delim_re
+from django.utils.datastructures import MultiValueDict
 from django.utils.http import http_date
+
+from . import http_client
 
 
 def reduce_vary_headers(response, additional):
@@ -70,3 +73,29 @@ def merge_fragment_cookies(response, fragment_cookies):
             # BaseCookie.__setitem__.
             dict.__setitem__(cookies, key, morsel)
     response.cookies = cookies
+
+def replace_esi_tags(request, response, urls):
+    if not urls:
+        return
+
+    fragment_headers = MultiValueDict()
+    fragment_cookies = []
+    request_data = {
+        'cookies': request.COOKIES,
+        'HTTP_REFERER': request.build_absolute_uri(),
+    }
+
+    for url in urls:
+        esi_tag = '<esi:include src="%s" />' % url
+        client = http_client.Client(**request_data)
+        fragment = client.get(url)
+        response.content = response.content.replace(esi_tag, fragment.content)
+
+        for header in HEADERS_TO_MERGE:
+            if header in fragment:
+                fragment_headers.appendlist(header, fragment[header])
+        if fragment.cookies:
+            fragment_cookies.append(fragment.cookies)
+
+    merge_fragment_headers(response, fragment_headers)
+    merge_fragment_cookies(response, fragment_cookies)
